+++ src/components/pageBroteApp/index.ts
@@ -1,0 +1,14 @@
+import PageOpenCal from './pageBroteApp';
+import { register } from '../../shared/functions';
+
+import '../componentGroupButton/index';
+import '../pageCodeScanner/index';
+import '../pageSearch/index';
+import '../pageFood/index';
+import '../pageUser/index';
+import '../pageHome/index';
+import '../pageHome/index';
+import '../pageMeal/index';
+
+
+register('brote-app', PageOpenCal);